Updating data for March 29, 2020 due to data source change from MOH.
Update March 30, 2020: Ministry of Health has updated the source data they use to report cases. As a result, there was a slight revision to the numbers reported for March 29, 2020. The biggest change is to the number of positive case (going up from 1324 to 1706), cases resolved (going up from 8 to 431) as well as the number of cases under investigation (going down from 7203 to 5651). There was also a slight revision to the number of test conducted (revised downward by about 1000 cases) and number of negative cases (going up by 20).